+++ src/rune-bridge/src/ledger.rs
@@ -1,53 +1,27 @@
-use std::borrow::Cow;
+mod used_utxo_details;
+mod utxo_details;
+mod utxo_key;
+mod utxo_runes;
+
 use std::collections::HashMap;
-use std::fmt;
-use std::mem::size_of;
-use std::str::FromStr;
 
 use bitcoin::hashes::sha256d::Hash;
-use bitcoin::{Address, Amount, Network, OutPoint, TxOut, Txid};
-use candid::{CandidType, Decode, Encode};
-use ic_exports::ic_cdk::api::management_canister::bitcoin::{Outpoint, Utxo};
+use bitcoin::{Address, Amount, OutPoint, TxOut, Txid};
+use ic_exports::ic_cdk::api::management_canister::bitcoin::Utxo;
 use ic_exports::ic_kit::ic;
 use ic_stable_structures::stable_structures::DefaultMemoryImpl;
-use ic_stable_structures::{BTreeMapStructure, Bound, StableBTreeMap, Storable, VirtualMemory};
+use ic_stable_structures::{BTreeMapStructure, StableBTreeMap, VirtualMemory};
 use ord_rs::wallet::TxInputInfo;
-use serde::Deserialize;
 
-use crate::key::{ic_dp_to_derivation_path, IcBtcSigner};
+use self::used_utxo_details::UsedUtxoDetails;
+use self::utxo_details::UtxoDetails;
+pub use self::utxo_key::UtxoKey;
+use self::utxo_runes::UtxoRunes;
+use crate::key::ic_dp_to_derivation_path;
 use crate::memory::{
     LEDGER_MEMORY_ID, MEMORY_MANAGER, RUNE_INFO_BY_UTXO_MEMORY_ID, USED_UTXOS_REGISTRY_MEMORY_ID,
 };
 use crate::rune_info::RuneInfo;
-
-/// Data structure to keep track rune information for a utxo.
-struct UtxoRunes(Vec<RuneInfo>);
-
-impl Storable for UtxoRunes {
-    const BOUND: Bound = Bound::Unbounded;
-
-    fn from_bytes(bytes: Cow<[u8]>) -> Self {
-        let list_len = u64::from_be_bytes(bytes[0..8].try_into().unwrap()) as usize;
-        let mut runes = Vec::with_capacity(list_len);
-
-        for rune_buf in bytes[8..bytes.len()].chunks(RuneInfo::BOUND.max_size() as usize) {
-            runes.push(RuneInfo::from_bytes(Cow::Borrowed(rune_buf)));
-        }
-
-        Self(runes)
-    }
-
-    fn to_bytes(&self) -> Cow<[u8]> {
-        let mut bytes = Vec::with_capacity(8 + self.0.len() * RuneInfo::BOUND.max_size() as usize);
-        bytes.extend_from_slice(&(self.0.len() as u64).to_be_bytes());
-
-        for rune in &self.0 {
-            bytes.extend_from_slice(&rune.to_bytes());
-        }
-
-        Cow::Owned(bytes)
-    }
-}
 
 /// Data structure to keep track of utxos owned by the canister.
 pub struct UtxoLedger {
@@ -70,113 +44,6 @@ impl Default for UtxoLedger {
             ),
         }
     }
-}
-
-#[derive(Debug, Copy, Clone, Eq, PartialEq, Ord, PartialOrd, CandidType, Deserialize, Hash)]
-pub struct UtxoKey {
-    pub tx_id: [u8; 32],
-    pub vout: u32,
-}
-
-impl fmt::Display for UtxoKey {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}:{}", hex::encode(self.tx_id), self.vout)
-    }
-}
-
-impl Storable for UtxoKey {
-    fn to_bytes(&self) -> Cow<[u8]> {
-        let bytes = Encode!(self).expect("cannot serialize utxo key");
-        Cow::Owned(bytes)
-    }
-
-    fn from_bytes(bytes: Cow<[u8]>) -> Self {
-        Decode!(&bytes, Self).expect("cannot deserialize utxo key")
-    }
-
-    const BOUND: Bound = Bound::Bounded {
-        max_size: 60,
-        is_fixed_size: true,
-    };
-}
-
-impl From<OutPoint> for UtxoKey {
-    fn from(value: OutPoint) -> Self {
-        Self {
-            tx_id: *<Hash as AsRef<[u8; 32]>>::as_ref(&value.txid.to_raw_hash()),
-            vout: value.vout,
-        }
-    }
-}
-
-impl From<&Outpoint> for UtxoKey {
-    fn from(value: &Outpoint) -> Self {
-        Self {
-            tx_id: value.txid.clone().try_into().expect("invalid tx id"),
-            vout: value.vout,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Eq, PartialEq, CandidType, Deserialize)]
-pub struct UtxoDetails {
-    value: u64,
-    script_buf: Vec<u8>,
-    derivation_path: Vec<Vec<u8>>,
-}
-
-impl UtxoDetails {
-    const MAX_SCRIPT_SIZE: u32 = 128;
-    const DERIVATION_PATH_SIZE: u32 = IcBtcSigner::DERIVATION_PATH_SIZE;
-}
-
-impl Storable for UtxoDetails {
-    fn to_bytes(&self) -> Cow<[u8]> {
-        let bytes = Encode!(self).expect("failed to serialize utxo");
-        Cow::Owned(bytes)
-    }
-
-    fn from_bytes(bytes: Cow<[u8]>) -> Self {
-        Decode!(&bytes, Self).expect("failed to deserialize utxo")
-    }
-
-    const BOUND: Bound = Bound::Bounded {
-        max_size: size_of::<u64>() as u32 + Self::MAX_SCRIPT_SIZE + Self::DERIVATION_PATH_SIZE,
-        is_fixed_size: false,
-    };
-}
-
-#[derive(Debug, Clone, Eq, PartialEq, CandidType, Deserialize)]
-pub struct UsedUtxoDetails {
-    pub used_at: u64,
-    owner_address: String,
-}
-
-impl UsedUtxoDetails {
-    const MAX_BITCOIN_ADDRESS_SIZE: u32 = 96;
-
-    /// Returns the owner address of the utxo.
-    pub fn owner_address(&self, network: Network) -> Result<Address, Box<dyn std::error::Error>> {
-        Address::from_str(self.owner_address.as_str())?
-            .require_network(network)
-            .map_err(|e| e.into())
-    }
-}
-
-impl Storable for UsedUtxoDetails {
-    fn to_bytes(&self) -> Cow<[u8]> {
-        let bytes = Encode!(self).expect("failed to serialize utxo");
-        Cow::Owned(bytes)
-    }
-
-    fn from_bytes(bytes: Cow<[u8]>) -> Self {
-        Decode!(&bytes, Self).expect("failed to deserialize utxo")
-    }
-
-    const BOUND: Bound = Bound::Bounded {
-        max_size: size_of::<u64>() as u32 + Self::MAX_BITCOIN_ADDRESS_SIZE,
-        is_fixed_size: false,
-    };
 }
 
 /// Information about the unspent utxo.
@@ -291,86 +158,13 @@ impl UtxoLedger {
 mod tests {
     use std::str::FromStr;
 
-    use bitcoin::{Network, PublicKey};
-    use did::H160;
+    use ic_exports::ic_cdk::api::management_canister::bitcoin::Outpoint;
     use ic_exports::ic_kit::MockContext;
     use ordinals::Rune;
 
     use super::*;
     use crate::canister::get_rune_state;
-    use crate::key::get_derivation_path_ic;
     use crate::rune_info::RuneName;
-
-    #[test]
-    fn key_serialization() {
-        let key = UtxoKey {
-            tx_id: [123; 32],
-            vout: 544331,
-        };
-
-        let serialized = key.to_bytes();
-        let Bound::Bounded { max_size, .. } = UtxoKey::BOUND else {
-            panic!("Key is unbounded");
-        };
-
-        assert_eq!(serialized.len() as u32, max_size);
-        let deserialized = UtxoKey::from_bytes(serialized);
-        assert_eq!(deserialized, key);
-    }
-
-    #[test]
-    fn value_serialization() {
-        let address = Address::p2wpkh(
-            &PublicKey::from_str(
-                "038f47dcd43ba6d97fc9ed2e3bba09b175a45fac55f0683e8cf771e8ced4572354",
-            )
-            .unwrap(),
-            Network::Signet,
-        )
-        .unwrap();
-        let derivation_path = get_derivation_path_ic(
-            &H160::from_hex_str("0x0dc9f6938e9b47fd8553df50bcbdb62d67239007").unwrap(),
-        );
-        let value = UtxoDetails {
-            value: 100500,
-            script_buf: address.script_pubkey().to_bytes(),
-            derivation_path,
-        };
-
-        let serialized = value.to_bytes();
-        let Bound::Bounded { max_size, .. } = UtxoDetails::BOUND else {
-            panic!("Key is unbounded");
-        };
-
-        assert!((serialized.len() as u32) < max_size);
-        let deserialized = UtxoDetails::from_bytes(serialized);
-        assert_eq!(deserialized, value);
-    }
-
-    #[test]
-    fn test_should_serialize_used_utxo() {
-        let address = Address::p2wpkh(
-            &PublicKey::from_str(
-                "038f47dcd43ba6d97fc9ed2e3bba09b175a45fac55f0683e8cf771e8ced4572354",
-            )
-            .unwrap(),
-            Network::Signet,
-        )
-        .unwrap();
-        let value = UsedUtxoDetails {
-            used_at: 100500,
-            owner_address: address.to_string(),
-        };
-
-        let serialized = value.to_bytes();
-        let Bound::Bounded { max_size, .. } = UsedUtxoDetails::BOUND else {
-            panic!("Key is unbounded");
-        };
-
-        assert!((serialized.len() as u32) < max_size);
-        let deserialized = UsedUtxoDetails::from_bytes(serialized);
-        assert_eq!(deserialized, value);
-    }
 
     #[test]
     fn test_should_deposit_utxo() {
